refactor: rename workbox-config.js to workbox.config.js
