+++ .gitignore
@@ -18,9 +18,7 @@ Desktop.ini
 *.map
 *.d.ts
 *.js
-!/*.js
-!*config.js
-!esbuild.config.js
+!*.config.js
 dist
 build
